Add Fearsome Bark ability to Gnolls
Adds the Fearsome Bark ability to the Gnoll species. Using this ability
attempts to fear all enemies in line of sight (currently at the same
strength as a scroll of fear), while making a loud noise at the player's
location (25, same as a scroll of noise). It can be used even if the
player is confused.

This ability is currently granted at XL 1. This is just a placeholder, and
can be changed to a higher level for balance reasons.

Placeholder ability tile credit goes to Xellpnz if it isn't changed before
release.